🤖 Fix tests
