+++ frontend/components/forms/fields/InputFieldWithIcon/InputFieldWithIcon.tests.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { renderWithSetup } from "test/test-utils";
-// @ts-ignore
 import InputFieldWithIcon from "./InputFieldWithIcon";
 
 describe("InputFieldWithIcon Component", () => {
